requirements.txt
dnspython==2.3.0
duckdb==0.6.1
numpy==1.24.1
pymongo==4.3.3
python-dotenv==0.21.0
